% Generated by r6-generator-maven-plugin: do not edit by hand
\docType{package}
\name{jplantuml4r-package}
\alias{jplantuml4r}
\alias{jplantuml4r-package}
\title{jplantuml4r: Jplantuml4r: R6 Java Wrapper Package}
\description{
% \if{html}{\figure{logo.png}{options: align='right'}}
An R6 package wrapping java code in the org.github.io.github.terminological:jplantuml4r library. This library was generated by the r6-generator-maven-plugin.

Version: 0.0.0.9000

Classes:
\itemize{ 
	\item \code{JavaApi}
	\item \code{PlantUML}
}
}

\author{
\itemize{ 
   \item Rob Challen rob.challen@bristol.ac.uk 0000-0002-5504-7768
}
}

\keyword{java api}