% Generated by r6-generator-maven-plugin: do not edit by hand

\name{PlantUML}

\alias{PlantUML}

\title{PlantUML}



\description{
PlantUML can generate a variety of diagrams from a DSL

Version: 0.0.0.9000

Generated: 2024-04-24T15:09:41.095348424
}

\details{
	PlantUML can generate a variety of diagrams from a DSL
}

\examples{
## -----------------------------------
## Construct new instance of PlantUML
## -----------------------------------
\dontrun{
J = jplantuml4r::JavaApi$get()
# appropriate parameter values must be provided
instance = J$PlantUML$new()
}
}

\keyword{java api}

\section{Methods}{
	\subsection{Constructors}{
		\itemize{
			\item \href{#method-new}{\code{J$PlantUML$new()}}
		}
	}
	\subsection{Class methods}{
		\itemize{
			\item{none}
			\item \code{instance$clone()}
			\item \code{instance$print()}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Constructor \code{new()}}{
the default no-args constructor

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jplantuml4r::JavaApi$get()
instance = J$PlantUML$new();
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
the new R6 PlantUML object
		}

	}

}