% Generated by r6-generator-maven-plugin: do not edit by hand

\name{save_plant_uml}
\alias{save_plant_uml}
\title{Render a plantuml file to a png file}
\usage{save_plant_uml(
	plantuml,
	outFile,
	format
)
}
\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{plantuml}{plantuml a plantuml file as a character vector. - (java expects a RCharacterVector)}
	\item{outFile}{outFile the output filename - (defaulting to `tempfile()`) - (java expects a RFile)}
	\item{format}{format see supported_formats method - (defaulting to `'png'`) - (java expects a RCharacter)}
	\if{html}{\out{</div>}}
}
\value{
RFile: this java method returns the message that the object was created with

}
\description{
no description
}
\keyword{java api}


